# modules/nims/metadata-service/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NIM Metadata Service",
    "description": "Provides information about NVIDIA NIM deployments running in the cluster",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/nims/{nim_name}": {
      "get": {
        "tags": [
          "nims"
        ],
        "summary": "Get Nim",
        "description": "Get detailed information about a specific NIM deployment.\n\nArgs:\n    nim_name: Name of the NIM deployment (e.g., 'openfold3', 'boltz2')",
        "operationId": "get_nim_api_v1_nims__nim_name__get",
        "parameters": [
          {
            "name": "nim_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Nim Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NIMInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NIMInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "image": {
            "type": "string",
            "title": "Image"
          },
          "replicas_desired": {
            "type": "integer",
            "title": "Replicas Desired"
          },
          "replicas_ready": {
            "type": "integer",
            "title": "Replicas Ready"
          },
          "replicas_available": {
            "type": "integer",
            "title": "Replicas Available"
          },
          "port": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Port"
          },
          "resources": {
            "$ref": "#/components/schemas/ResourceInfo"
          },
          "pods": {
            "items": {
              "$ref": "#/components/schemas/PodInfo"
            },
            "type": "array",
            "title": "Pods"
          }
        },
        "type": "object",
        "required": [
          "name",
          "image",
          "replicas_desired",
          "replicas_ready",
          "replicas_available",
          "resources",
          "pods"
        ],
        "title": "NIMInfo"
      },
...
      "PodInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "node": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Node"
          },
          "gpu_index": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gpu Index"
          },
          "restart_count": {
            "type": "integer",
            "title": "Restart Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "name",
          "status"
        ],
        "title": "PodInfo"
      },
      "ResourceInfo": {
        "properties": {
          "cpu_request": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cpu Request"
          },
          "cpu_limit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cpu Limit"
          },
          "memory_request": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memory Request"
          },
          "memory_limit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memory Limit"
          },
          "gpu_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gpu Count"
          }
        },
        "type": "object",
        "title": "ResourceInfo"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}